courseapis/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
mysqlclient==2.2.7
pillow==11.1.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
